adicao dos services

+++ src/main/java/com/djg/repository/TransacaoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.djg.model.Transacao;
 
-public interface TransacaoRepository extends CrudRepository<Transacao, Integer>{
+public interface TransacaoRepository extends CrudRepository<Transacao, Long>{
 
 }
